fix: DSP correctness + graph optimizer passes + 21 op validation fixes
DSP fixes:
- argTableStable data pointer refresh for gradient checking (putScalar invisible to DSP)
- LEGACY_PAIRWISE_BOOL (18) for or/and/xor ops — DSP classifier checks opType() before instanceof
- LegacyPairwiseTransformBoolOp: execPairwiseBoolTransform instead of execPairwiseTransform
- Zero-slot passthrough for empty plans (pow(x,1)→x optimization)
- VARIABLE array syncToDevice before DSP execution
- DSP view guards for Permute/StridedSlice/Reshape/ReshapeNoCopy initializeOutputs
- Remove DSP disabling from GradCheckUtil — grad check must work with DSP enabled

Op fixes:
- lrelu_bp: cast() returns NDArray* (heap), must delete after use
- inTopKFunctor_: rewrite with count-greater approach, result->nullify()
- ClipByNorm/ClipByAvgNorm: SameDiff constructor not passing dimensions to iArguments
- Cast: empty scalar handling for rank-0 arrays (check input.isEmpty() directly)
- IsMax: null OpContext handling, BOOL output type
- clip.cpp/clip.cu: e<double>(0) instead of e<float>(0) for DOUBLE precision
- BroadcastableOp: empty broadcast returns scalar empty (TF semantics)
- BaseDynamicTransformOp: empty input → scalar empty output
- transpose.cpp: remove DOUBLE from oneDNN PLATFORM_CHECK (f64 unsupported)
- InTopK/Trace: calculateOutputShapeFromInputs implementations
- DynamicCustomOp: null guards in iArgs/tArgs/dArgs

Graph optimizer passes (new):
- CommonSubexpressionElimination, HorizontalFusion, StrengthReduction
- Rematerialization, BroadcastElimination, Reordering
- RedundancyElimination, Peephole, ArithmeticChain
- MatMulChain, SelectWhere, ConcatSplit
- AutoTuner (execution mode selector)
- ForwardExecutionDAGBuilder (memory-aware scheduling)
- NativePlanCompiler (buffer aliasing)
- GraphOptimizer.defaultCorrectnessOptimizations()

Test fixes for all 220 tests passing (0 failures, 0 errors).


Former-commit-id: b1e9e87f01b14aac77d94ee665924759e2145d47

+++ libnd4j/include/ops/declarable/helpers/cpu/clip.cpp
@@ -48,14 +48,14 @@ void clipByNorm(LaunchContext* context, NDArray* input, NDArray* output, const s
     NDArray *norm2Result = z->reduceAlongDimension(reduce::Norm2, &emptyVec);
     if (useAverage) {
       NDArray *divResult = (*norm2Result) / z->lengthOf();
-      if (divResult->e<float>(0) > clipNorm->e<float>(0)) {
+      if (divResult->e<double>(0) > clipNorm->e<double>(0)) {
         NDArray *clipDivResult = (*clipNorm) / (*divResult);
         *z *= (*clipDivResult);
         delete clipDivResult;
       }
       delete divResult;
     } else {
-      if (norm2Result->e<float>(0) > clipNorm->e<float>(0)) {
+      if (norm2Result->e<double>(0) > clipNorm->e<double>(0)) {
         NDArray *clipDivResult = (*clipNorm) / (*norm2Result);
         *z *= (*clipDivResult);
         delete clipDivResult;
@@ -71,14 +71,14 @@ void clipByNorm(LaunchContext* context, NDArray* input, NDArray* output, const s
         NDArray *norm2Result = listOfSubArrs.at(i)->reduceAlongDimension(reduce::Norm2, &emptyVec);
         if (useAverage) {
           NDArray *divResult = (*norm2Result) / listOfSubArrs.at(i)->lengthOf();
-          if (divResult->e<float>(0) > clipNorm->e<float>(0)) {
+          if (divResult->e<double>(0) > clipNorm->e<double>(0)) {
             NDArray *clipDivResult = (*clipNorm) / (*divResult);
             *listOfSubArrs.at(i) *= (*clipDivResult);
             delete clipDivResult;
           }
           delete divResult;
         } else {
-          if (norm2Result->e<float>(0) > clipNorm->e<float>(0)) {
+          if (norm2Result->e<double>(0) > clipNorm->e<double>(0)) {
             NDArray *clipDivResult = (*clipNorm) / (*norm2Result);
             *listOfSubArrs.at(i) *= (*clipDivResult);
             delete clipDivResult;
